finished #18
Unlinked fb acc is redirected to register page. Email field is filled with email from facebook. After registering account will be linked with facebook.

If user has created account with same email as in Facebook, then that account is linked with Facebook and user is logged in with that account.